requirements.txt
playwright==1.41.0
paho-mqtt==1.6.1
schedule==1.2.1
websockets
python-dateutil
